+++ changelog.md
@@ -4,6 +4,8 @@
 * Keego
   * Mining Momentum now requires 24 stacks to be fully stacked instead of 32 (~~+10%~~ -> +15% mining speed per stack)
   * Increased minimum mining effect time (0.75 -> 1 seconds)
+  * Added infos in EMI
+* Vanilla blast furnace recipes are now hidden from EMI
 * Detector Rails no longer need redstone
 
 ## 1.9.6